+++ workspace/app/src/main/java/com/clilystudio/netbook/widget/comca/m.java
@@ -3,8 +3,8 @@ package com.clilystudio.netbook.widget.comca;
 public final class m implements v<Integer> {
     @Override
     public final /* synthetic */ Integer a(float f, Integer object, Integer object2) {
-        Integer n = (Integer) object;
-        Integer n2 = (Integer) object2;
+        Integer n = object;
+        Integer n2 = object2;
         int n3 = n;
         return (int) ((float) n3 + f * (float) (n2 - n3));
     }
